requirements.txt
pygame
nuitka
requests
Pillow
PyQt6
